+++ x/interchange/types/orderbook_test.go
@@ -8,19 +8,19 @@ import (
 	"testing"
 )
 
-func GenAmount() uint64 {
-	return uint64(rand.Intn(int(types.MaxAmount)) + 1)
+func GenAmount() uint32 {
+	return uint32(rand.Intn(int(types.MaxAmount)) + 1)
 }
 
-func GenPrice() uint64 {
-	return uint64(rand.Intn(int(types.MaxPrice)) + 1)
+func GenPrice() uint32 {
+	return uint32(rand.Intn(int(types.MaxPrice)) + 1)
 }
 
 func GenPair() (string, string) {
 	return GenString(10), GenString(10)
 }
 
-func GenOrder() (types.Account, uint64, uint64) {
+func GenOrder() (types.Account, uint32, uint32) {
 	return GenLocalAccount(), GenAmount(), GenPrice()
 }
